+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.5.0 (TBD)
 
+* [BREAKING] Track expected block height for notes (#448).
 * [BREAKING] Updated Rust Client to use the new version of `miden-base` (#492).
 * [BREAKING] Removed `TransactionTemplate` and `account_id` from `TransactionRequest` (#478).
 * Added validation for consumed notes when importing (#449).